+++ lib/src/public_key/rsa.dart
@@ -3,6 +3,12 @@ library dartcrypto.ciphers.rsa;
 import 'dart:math' as math;
 
 class RSACipher {
+  int modulus, publicExponent, privateExponent, prime1, prime2, exponent1, exponent2, coefficient;
+
   RSACipher();
+
+  generatePEM() {
+    // TODO: implement;
+  }
   // TODO: implement;
 }